# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Life RPG API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /stats/{statName}/xp:
    post:
      operationId: addStatXp
      tags: [stats]
      summary: Add XP to a specific stat
      parameters:
        - name: statName
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddXpBody"
      responses:
        "200":
          description: Updated stat
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Stat"
...
  /quests/daily/{id}/complete:
    post:
      operationId: completeDailyQuest
      tags: [quests]
      summary: Mark a daily quest as complete
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Completed quest
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyQuest"
...
  /quests/side:
    get:
      operationId: getSideQuests
      tags: [quests]
      summary: Get all side quests
      responses:
        "200":
          description: Side quests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SideQuest"
    post:
      operationId: createSideQuest
      tags: [quests]
      summary: Create a side quest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSideQuestBody"
      responses:
        "201":
          description: Created side quest
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SideQuest"
  /quests/side/{id}/complete:
    post:
      operationId: completeSideQuest
      tags: [quests]
      summary: Complete a side quest
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Completed side quest
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SideQuest"
  /quests/main:
    get:
      operationId: getMainQuests
      tags: [quests]
      summary: Get all main quests
      responses:
        "200":
          description: Main quests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MainQuest"
    post:
      operationId: createMainQuest
      tags: [quests]
      summary: Create a main quest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMainQuestBody"
      responses:
        "201":
          description: Created main quest
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MainQuest"
  /quests/main/{id}/complete:
    post:
      operationId: completeMainQuest
      tags: [quests]
      summary: Complete a main quest
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Completed main quest
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MainQuest"
...
  /rewards/{id}/use:
    post:
      operationId: useReward
      tags: [rewards]
      summary: Use a reward from inventory
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reward used
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reward"
...
components:
  schemas:
...
    Stat:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        displayName:
          type: string
        xp:
          type: integer
        level:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, name, displayName, xp, level, title, createdAt, updatedAt]
    AddXpBody:
      type: object
      properties:
        amount:
          type: integer
        reason:
          type: string
      required: [amount]
...
    DailyQuest:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        xpReward:
          type: integer
        statReward:
          type: string
        completed:
          type: boolean
        date:
          type: string
          format: date-time
        completedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
      required: [id, title, description, xpReward, statReward, completed, date, createdAt]
    SideQuest:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        xpReward:
          type: integer
        statReward:
          type: string
        completed:
          type: boolean
        completedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
      required: [id, title, description, xpReward, statReward, completed, createdAt]
    CreateSideQuestBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        xpReward:
          type: integer
        statReward:
          type: string
      required: [title, description, xpReward, statReward]
    MainQuest:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        xpReward:
          type: integer
        completed:
          type: boolean
        completedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
      required: [id, title, description, xpReward, completed, createdAt]
    CreateMainQuestBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        xpReward:
          type: integer
      required: [title, description, xpReward]
...
    Reward:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
        name:
          type: string
        description:
          type: string
        used:
          type: boolean
        usedAt:
          type: ["string", "null"]
          format: date-time
        expiresAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
      required: [id, type, name, description, used, createdAt]
...