# lib/api-zod/src/generated/types/mainQuest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Life RPG API
 * OpenAPI spec version: 0.1.0
 */

export interface MainQuest {
  id: number;
  title: string;
  description: string;
  xpReward: number;
  completed: boolean;
  /** @nullable */
  completedAt?: Date | null;
  createdAt: Date;
}