# lib/api-zod/src/generated/types/stat.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Life RPG API
 * OpenAPI spec version: 0.1.0
 */

export interface Stat {
  id: number;
  name: string;
  displayName: string;
  xp: number;
  level: number;
  title: string;
  createdAt: Date;
  updatedAt: Date;
}